add traffic generator client for pod
For some test suites, such as BGP, it is necessary
to use a traffic generator inside the pod, rather than
in a virtual machine. This commit adds a PodClient class
that represents the tcp client running in the pod.

Also this commit adds "--connect-timeout 300" option to
iperf3 cmd to fix incorrect behavior and prevent false
positives if there is no direct connectivity between
the server and the client.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>