net,infra: add fixtures and functions for BGP setup (#1776)
* move up network_operator fixture

This fixture is general for the network tests and might
be useful not only for the flat_overlay test suite. For example,
network_operator fixture is required to set up the BGP suite.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>

* extend functionality of cudn, nncp, nad classes

New functionality:
1) ClusterUserDefinedNetwork: the ability to create Layer2 CUDNs;
2) NetworkAttachmentDefinition: class CNIPluginMacvlanConfig
to create the macvlan NAD type;
3) NodeNetworkConfigurationPolicy: class Vlan to create vlan
interfaces on the cluster nodes;
4) new function for udn\cudn NS creation;

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>

* add fixtures and lib functions for BGP setup

Added fixtures and functions:
1) to configure an external FRR router on one
of the cluster nodes;
2) to create a CUDN;
3) to create cluster resources such as routeAdvertisements
and FRRConfiguration for BGP testing.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>

* add _verify_bgp_env_vars() to network_sanity()

Before the tests run, it is supposed to be checked that
the cluster supports BGP tests and the necessary ENVs are set.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>

* initial version of README.md

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>

* quarantine bgp test

BGP test suite infra dependencies are not met,
tracked in CNV-69734

fix

---------

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>