net,bgp: improve sanity check (#2175)
We don't want to collect any BGP tests and check for the network tests
sanity while the infra is not ready for BGP. Test quarantining is not
enough: _verify_bgp_env_vars() is run anyway and check for ENVs,
and this might fail the whole test run. We need to prevent it.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>