tests, net, bgp: add connectivity test suite (#2121)
* move udn_vm function to libs

udn_vm is a useful function and
should be placed in a more common location.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>

* add traffic generator client for pod

For some test suites, such as BGP, it is necessary
to use a traffic generator inside the pod, rather than
in a virtual machine. This commit adds a PodClient class
that represents the tcp client running in the pod.

Also this commit adds "--connect-timeout 300" option to
iperf3 cmd to fix incorrect behavior and prevent false
positives if there is no direct connectivity between
the server and the client.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>

* add bgp connectivity test suite

Initial BGP testing includes checking the tcp connection between
the CUDN VM and the external network, as well as this connection
preservation during VM migration.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>

* quarantine tests

while CNV-69734 is open, tests should be quarantined

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>

---------

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>