net, traffic_generator: improve naming + retry decorator using
Client\Server looks quite common and such naming does not reflect
the real intention. The usage traffic_generator.Client does not
solve the problem as well. It's crucial to explicitly point out that
we work with Tcp traffic and the classes implement TCP client and
TCP server in reality. And this clarity is significant for tests also.

The commit also includes replacing TimeoutSampler with retry decorator,
that leads a little refactoring for classes to align the logic.
Retry-enabled `_ensure_is_running()` method is added to `__init__()`
to improve robustness and align the classes behaviour.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>